refactor: remove side field — long-only cleanup
- Remove Side type alias and side field from TradeCard dataclass
- Remove side parameter from build_trade_card, has_valid_setup, score_confluence
- Delete all short-side code branches (stops, TPs, trailing, thesis)
- Remove side filter from dashboard UI (toggle, badges, chips)
- Hardcode long-only in alert_store, trade_tracker, close_hold_scanner
- Update all tests to match new signatures
- Short records deleted from Supabase via SQL

+++ tests/test_alert_store_weekend.py
@@ -20,7 +20,7 @@ def test_save_alert_skips_weekend(monkeypatch):
     alert = {"trade_date": "2026-03-14", "symbol": "WEEKEND", "side": "long"}
     save_alert(alert)
     # Sunday
-    alert = {"trade_date": "2026-03-15", "symbol": "WEEKEND", "side": "short"}
+    alert = {"trade_date": "2026-03-15", "symbol": "WEEKEND", "side": "long"}
     save_alert(alert)
     # Weekday
     alert = {"trade_date": "2026-03-17", "symbol": "WEEKDAY", "side": "long"}
